mengurutkan product berdasarkan nama categories

+++ src/models/user.model.js
@@ -26,8 +26,12 @@ exports.findAll = async (searchKey='', sortBy="id", order="ASC", page=1) => {
         ORDER BY ${columnSort.join(', ')}
         LIMIT ${limit} OFFSET ${offset}
         `
+        console.log(sql)
         const values = [`%${searchKey}%`]
         const {rows} = await db.query(sql, values)
+        if(!rows.length){
+            throw new Error(`no data found`)
+        }
         return rows
     }
 
@@ -40,6 +44,9 @@ exports.findAll = async (searchKey='', sortBy="id", order="ASC", page=1) => {
     console.log(sql)
     const values = [`%${searchKey}%`]
     const {rows} = await db.query(sql, values)
+    if(!rows.length){
+        throw new Error(`no data found`)
+    }
     return rows
 }
 
@@ -70,31 +77,30 @@ exports.findOneByEmail = async (email) => {
 
 
 exports.insert = async (body) => {
-    console.log(body.password)
-    let role = ''
-    if(body.email === "admin@example.com" && body.password === "@adminPass"){
-        role = "admin"
-    }else if(body.email.includes('staff') && body.password === "@staffPass"){
-        role = "staff"
-    }else{
-        role = "customer"
-    }
-
     const queryString = await isStringExist("users", "email", body.email)                                        // melakukan query terlebih dahulu sebelum memasukan data, untuk mengecek apakah ada data string yg sama tapi hanya berbeda huruf kecil dan huruf besarnya saja. 
     if(queryString){
         throw new Error(queryString)
     }
 
+    // let role = ''
+    // if(body.email === "admin@example.com" && body.password === "@adminPass"){
+    //     role = "admin"
+    // }else if(body.email.includes('staff') && body.password === "@staffPass"){
+    //     role = "staff"
+    // }else{
+    //     role = "customer"
+    // }
+
     body.password = await argon.hash(body.password)
 
     const sql = `
     INSERT INTO "users"
-    ("fullName", "email", "password", "role")
+    ("fullName", "email", "password", "address", "picture", "phoneNumber", "role")
     VALUES
-    ($1, $2, $3, $4)
+    ($1, $2, $3, $4, $5, $6, $7)
     RETURNING *
     `
-    const values = [body.fullName, body.email, body.password, role]
+    const values = [body.fullName, body.email, body.password, body.address, body.picture, body.phoneNumber, body.role]
     const {rows} = await db.query(sql, values)
     return rows[0]
 }
@@ -117,11 +123,16 @@ exports.update = async (id, body) => {
 }
 
 
-exports.delete = async (id) => {                                                                                   // melakukan query terlebih dahulu sebelum delete, untuk mengecek apakah data yg ingin di delete ada di database
-    const queryId = await isExist("users", id)
+exports.delete = async (id) => {
+    if(isNaN(id)){                                                                            // error code 22P02 tidak ternotice karena ada isExist jadi lempar error custom jika data yg di masukan bukan number 
+        throw new Error(`invalid input`)                            
+    }            
+
+    const queryId = await isExist("users", id)                                                 // melakukan query terlebih dahulu sebelum delete, untuk mengecek apakah data yg ingin di delete ada di database
     if(queryId){
         throw new Error(queryId)
     }
+
     const sql = `DELETE FROM "users" WHERE "id" = $1 RETURNING *`
     const values = [id]
     const {rows} = await db.query(sql, values)
